% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\name{as_annotation}
\alias{as_annotation}
\title{Add annotation for highlighting to subcorpus.}
\usage{
as_annotation(x)
}
\arguments{
\item{x}{A \code{subcorpus} object.}
}
\description{
Add annotation for highlighting to subcorpus.
}
