% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{as_chunks}
\alias{as_chunks}
\title{Split vector into equally sized chunks}
\usage{
as_chunks(x, size)
}
\arguments{
\item{x}{A vector to be split into equally sized chunks.}

\item{size}{Numeric value, maximum size of chunks.}
}
\value{
A list of vectors.
}
\description{
Split input vector \code{x} into equally sized chunks. The maximum size is taken
from argument \code{size}.
}
