% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_subcorpus}
\alias{as_subcorpus}
\title{Transform table with DBpedia URIs to subcorpus.}
\usage{
as_subcorpus(x)
}
\arguments{
\item{x}{A \code{data.table} with DBpedia URIs.}
}
\description{
Transform table with DBpedia URIs to subcorpus.
}
