% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{dbpedia_get_wikidata_uris}
\alias{dbpedia_get_wikidata_uris}
\title{Get Wikipedia IDs for DBpedia IDS.}
\usage{
dbpedia_get_wikidata_uris(
  x,
  optional,
  endpoint,
  limit = 100,
  wait = 1,
  progress = FALSE
)
}
\arguments{
\item{x}{A character vector with DBpedia URIs.}

\item{optional}{Optional information to retrieve (passed as length-one
character vector, e.g. 'municipalityCode').}

\item{endpoint}{Endpoint to query (a \code{character} vector).}

\item{limit}{Single numeric value with maximum size of chunks to process at
a time.}

\item{wait}{A numeric value passed into \code{Sys.sleep()} to slow down sequence
of requests (and avoid denial of service). Defaults to 100.}

\item{progress}{Whether to show progress bar (\code{logical} value).}
}
\description{
Get Wikipedia IDs for DBpedia IDS.
}
\examples{
\donttest{
dbpedia_ids <- c(
  "http://de.dbpedia.org/resource/Killesberg",
  "http://de.dbpedia.org/resource/Ljubljana",
  "http://de.dbpedia.org/resource/Velbert"
)
dbpedia_get_wikidata_uris(
  dbpedia_ids,
  optional = "municipalityCode",
  endpoint = "http://de.dbpedia.org/sparql",
  wait = 0,
  limit = 2,
  progress = TRUE
)
}
}
