% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\name{dbpedia_lookup}
\alias{dbpedia_lookup}
\title{Use DBpedia Lookup service to get DBpedia URIs}
\usage{
dbpedia_lookup(
  query,
  api = "https://lookup.dbpedia.org/api/search",
  max_results = 5,
  progress = TRUE
)
}
\arguments{
\item{query}{A term/query to look up.}

\item{api}{The service to consult.}

\item{max_results}{The maximum number of results}

\item{progress}{A \code{logical} value, whether to show progress message.}
}
\value{
A named list of vectors with length \code{max_results}; the names are the
queries.
}
\description{
DBpedia Lookup is a service to match a query on a DBpedia URI. See the
explanation \href{https://www.dbpedia.org/resources/lookup/}{here} and the
experimental GUI \href{https://lookup.dbpedia.org/index.html}{here}.
}
\examples{
uri <- dbpedia_lookup(query = "Berlin")
uris <- dbpedia_lookup(query = c("Berlin", "Paris", "London"))
}
