% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\name{dbpedia_spotlight_status}
\alias{dbpedia_spotlight_status}
\alias{print.dbpedia_spotlight_status}
\title{Set and report status of DBpedia Spotlight}
\usage{
dbpedia_spotlight_status()

\method{print}{dbpedia_spotlight_status}(x, ...)
}
\arguments{
\item{x}{\code{dbpedia_spotlight_status} object to be printed.}

\item{...}{Further arguments passed into \code{print()}. Unused / required for
technical reasons.}
}
\value{
Object of class \code{dbpedia_spotlight_stats}, a \code{list} with elements
"docker" (\code{TRUE}/\code{FALSE}), "lang" and "endpoint".
}
\description{
Check whether Docker container with DBpedia Spotlight is running locally
and set options 'dbpedia.lang' and 'dbpedia.endpoint' accordingly.
}
\examples{
dbpedia_spotlight_status()
getOption("dbpedia.endpoint")
getOption("dbpedia.lang")
}
