% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\docType{data}
\name{dbpedia_stopwords}
\alias{dbpedia_stopwords}
\title{Stopwords used by DBpedia Spotlight}
\format{
An object of class \code{list} of length 24.
}
\usage{
dbpedia_stopwords
}
\description{
\code{dbpedia_stopwords} is a list of character vectors with stopwords used by
DBpedia Spotlight before processing chunks of texts. The data is used for
mapping offset positions returned from DBpedia Spotlight on corpus positions
on the R side.
}
\details{
\code{dbpedia_stopwords} is prepared using the script 'stopwords.R' in the
'data-raw' folder of the GitHub repository. The original data is included
in the \href{https://github.com/dbpedia-spotlight/model-quickstarter}{model-quickstarter}
repository of the DBpedia Spotlight project.
}
\examples{
names(dbpedia_stopwords) # languages available
}
\keyword{datasets}
