% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{detect_overlap}
\alias{detect_overlap}
\title{Detect overlapping entities}
\usage{
detect_overlap(x, start_col, end_col = NULL, verbose = TRUE)
}
\arguments{
\item{x}{The input \code{data.table} containing a start and an end position of
regions to be compared.}

\item{start_col}{A \verb{character vector} of length 1, referring to the name of
the column containing the start positions of the regions to be compared.}

\item{end_col}{A \verb{character vector} of length 1, referring to the name of the
column containing the end positions of the regions to be compared.}

\item{verbose}{A \verb{logical value} of whether to print messages or not.}
}
\value{
The input \code{data.table} is modified by reference. The column "ovl_id"
(overlap ID) is added. Each group of overlapping entities is annotated with
a unique ID. The functions \code{categorize_overlap()} and \code{resolve_overlap()}
can be used thereafter.
}
\description{
\code{detect_overlap()} detects overlapping regions of annotations in a \code{data.table}
object and assigns an overlap id to each group of overlapping entities.
\strong{This functionality is currently experimental}. The correctness of results
is subject to ongoing checks. The syntax is likely to change in future
updates.
}
\examples{
doc <- "Der Deutsche Bundestag tagt in Berlin."

x <- get_dbpedia_uris(
  x = doc,
  language = "de",
  max_len = 5600L,
  confidence = 0.35,
  api = "https://api.dbpedia-spotlight.org/de/annotate",
  types = character(),
  support = 20,
  types_src = c("DBpedia", "Wikidata"),
  verbose = TRUE
) |>
  detect_overlap(start_col = "start", verbose = TRUE)
}
