% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_table.R
\name{get_annotation_table}
\alias{get_annotation_table}
\alias{get_annotation_table,subcorpus-method}
\alias{get_annotation_table,subcorpus_bundle-method}
\title{Turn information on annotations to table.}
\usage{
get_annotation_table(x)

\S4method{get_annotation_table}{subcorpus}(x)

\S4method{get_annotation_table}{subcorpus_bundle}(x)
}
\arguments{
\item{x}{A \code{subcorpus} or \code{subcorpus_bundle} object.}
}
\description{
Turn information on annotations to table.
}
