% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href.R
\name{href}
\alias{href}
\title{Add hyperlinks to XML/HTML document.}
\usage{
href(x, url = list())
}
\arguments{
\item{x}{Input HTML object}

\item{url}{A named list.}
}
\description{
Add hyperlinks to XML/HTML document.
}
