% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_types_to_class}
\alias{map_types_to_class}
\title{Map types returned by DBpedia Spotlight to a limited set of classes}
\usage{
map_types_to_class(x, mapping_vector, other = "MISC", verbose = TRUE)
}
\arguments{
\item{x}{A \code{data.table} with DBpedia URIs.}

\item{mapping_vector}{A \verb{named character vector} with desired class names
(as names) and types from the DBpedia ontology as values.}

\item{other}{a \verb{character vector} with the name of the class of all types
not matched by the \code{mapping_vector}.}

\item{verbose}{A \code{logical} value - whether to display messages.}
}
\description{
Map types returned by DBpedia Spotlight to a limited set of classes
}
