% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{namespaced_xpath}
\alias{namespaced_xpath}
\title{function which adds the namespace to the XPATH}
\usage{
namespaced_xpath(xml = xml, tags)
}
\arguments{
\item{xml}{xml the namespace is derived from}

\item{tags}{the tags which should be queried with the XPATH}
}
\description{
function which adds the namespace to the XPATH
}
