% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\name{to_annotation}
\alias{to_annotation}
\title{Transform XML to NLP::AnnotatedPlainTextDocument}
\usage{
to_annotation(nodes, xml, token_tags, feature_tag)
}
\arguments{
\item{nodes}{which nodes constitute a single document, i.e. either a
nodeset of different segments. Can also be an entire XML document.}

\item{xml}{the xml document itself - needed to extract the namespace}

\item{token_tags}{...}

\item{feature_tag}{...}
}
\description{
Transform XML to NLP::AnnotatedPlainTextDocument
}
