% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wikidata_query}
\alias{wikidata_query}
\alias{wikidata_query,character-method}
\alias{wikidata_query,data.table-method}
\title{Query Wikidata endpoint for additional information.}
\usage{
wikidata_query(x, ...)

\S4method{wikidata_query}{character}(
  x,
  id,
  language = getOption("dbpedia.lang"),
  limit = 100L,
  wait = 1,
  verbose = TRUE,
  progress = FALSE
)

\S4method{wikidata_query}{data.table}(
  x,
  id,
  language = getOption("dbpedia.lang"),
  limit = 100L,
  wait = 1,
  verbose = TRUE,
  progress = FALSE
)
}
\arguments{
\item{x}{A vector of wikidata ids.  NA values are dropped, only unique values
will be processed.}

\item{...}{Further arguments.}

\item{id}{Wikidata ID for information to retrieve (\code{character} vector).}

\item{language}{Language to be used in Wikidata query.}

\item{limit}{Maximum number of wikidata IDs to be sent to endpoint at a time.}

\item{wait}{A numeric value - slow down requests to avoid denial of service.}

\item{verbose}{Whether to output messages (\code{logical} value).}

\item{progress}{Whether to show progress information (\code{logical} value).}
}
\value{
A \code{tibble}.
}
\description{
This is a wrapper for \code{WikidataQueryServiceR::query_wikidata()} to get
additional information for known wikidata IDs.
}
\examples{
wikidata_ids <- c("Q1741365", "Q3840", "Q437")
wikidata_query(wikidata_ids, id = "P439", progress = TRUE)

wikidata_uris <- c(
  "http://www.wikidata.org/entity/Q1020", 
  "http://www.wikidata.org/entity/Q27468",
  "http://www.wikidata.org/entity/Q2131751"
)
wikidata_query(wikidata_uris, id = "P439", progress = TRUE)
}
