% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\docType{data}
\name{dbpedia_stopwords}
\alias{dbpedia_stopwords}
\title{Stopwords used by DBpedia Spotlight}
\format{
An object of class \code{list} of length 24.
}
\usage{
dbpedia_stopwords
}
\description{
Stopwords used by DBpedia Spotlight
}
\keyword{datasets}
