% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{detect_overlap}
\alias{detect_overlap}
\title{Detect overlapping entities}
\usage{
detect_overlap(x, start_col, end_col = NULL, verbose = TRUE)
}
\arguments{
\item{x}{the input \code{data.table} containing a start and an end position of
regions to be compared.}

\item{start_col}{a \verb{character vector} of length 1, referring to the name of
the column containing the start positions of the regions to be compared.}

\item{end_col}{a \verb{character vector} of length 1, referring to the name of the
column containing the end positions of the regions to be compared.}

\item{verbose}{a \verb{logical value} of whether to print messages or not.}
}
\value{
The input \code{data.table} is modified by reference. The column "ovl_id"
(overlap ID) is added. Each group of overlapping entities is annotated with
a unique ID. The functions \code{classify_overlap()} and \code{resolve_overlap()} can
be used thereafter.
}
\description{
The function detects overlapping regions of annotations in a \code{data.table}
object and assigns an overlap id to each group of overlapping entities.
}
\examples{
doc <- "Der Deutsche Bundestag tagt in Berlin."

get_dbpedia_uris(
  x = doc,
  language = getOption("dbpedia.lang"),
  max_len = 5600L,
  confidence = 0.35,
  api = getOption("dbpedia.endpoint"),
  types = character(),
  support = 20,
  types_src = c("DBpedia", "Wikidata"),
  verbose = TRUE
) |>
  detect_overlap(start = "start", verbose = TRUE)
}
