% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\name{get_dbpedia_links}
\alias{get_dbpedia_links}
\title{Get DBpedia links.}
\usage{
get_dbpedia_links(
  x,
  language,
  max_len = 6067L,
  confidence = 0.35,
  api = "http://localhost:2222/rest/annotate",
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{subcorpus} object. Will be coerced to
'AnnotatedPlainTextDocument' from NLP package.}

\item{language}{The language of the input text ("en", "fr", "de", ...) to
determine the stopwords used.}

\item{max_len}{An \code{integer} value. The text passed to DBpedia Spotlight may
not exceed a defined length. If it does, an HTTP error results. The known
threshold of 6067 characters is the default value.}

\item{confidence}{A \code{numeric} value, the minimum similarity score that serves
as threshold befor DBpedia Spotlight includes a link into the report.}

\item{api}{An URL of the DBpedia Spotlight API.}

\item{verbose}{A \code{logical} value - whether to display progress messages.}
}
\description{
Get DBpedia links.
}
