% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia.R
\docType{methods}
\name{get_dbpedia_uris}
\alias{get_dbpedia_uris}
\alias{get_dbpedia_uris,AnnotatedPlainTextDocument-method}
\alias{get_dbpedia_uris,subcorpus-method}
\alias{get_dbpedia_uris,subcorpus_bundle-method}
\title{Get DBpedia links.}
\usage{
get_dbpedia_uris(x, ...)

\S4method{get_dbpedia_uris}{AnnotatedPlainTextDocument}(
  x,
  language,
  max_len = 6067L,
  confidence = 0.35,
  api = "http://localhost:2222/rest/annotate",
  verbose = TRUE
)

\S4method{get_dbpedia_uris}{subcorpus}(
  x,
  language,
  p_attribute = "word",
  s_attribute = NULL,
  max_len = 6067L,
  confidence = 0.35,
  api = "http://localhost:2222/rest/annotate",
  verbose = TRUE
)

\S4method{get_dbpedia_uris}{subcorpus_bundle}(
  x,
  language,
  p_attribute = "word",
  s_attribute = NULL,
  confidence = 0.35,
  api = "http://localhost:2222/rest/annotate",
  max_len = 6067L,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{subcorpus} object. Will be coerced to
'AnnotatedPlainTextDocument' from NLP package.}

\item{...}{Further arguments.}

\item{language}{The language of the input text ("en", "fr", "de", ...) to
determine the stopwords used.}

\item{max_len}{An \code{integer} value. The text passed to DBpedia Spotlight may
not exceed a defined length. If it does, an HTTP error results. The known
threshold of 6067 characters is the default value.}

\item{confidence}{A \code{numeric} value, the minimum similarity score that serves
as threshold befor DBpedia Spotlight includes a link into the report.}

\item{api}{An URL of the DBpedia Spotlight API.}

\item{verbose}{A \code{logical} value - whether to display progress messages.}

\item{p_attribute}{The p-attribute used for decoding a \code{subcorpus} object.}

\item{s_attribute}{A length-one \code{character} vector indicating a s-attribute.
DBpedia URIs will be mapped on this s-attribute. Only regions covered by
this s-attribute will be kept. If missing, URIs will be
mapped on the token stream, and all URIs suggested will be kept.}
}
\value{
A \code{data.table} with columns 'dbpedia_uri' and 'text'. Depending on
input object, further columns are ...
}
\description{
Get DBpedia links.
}
\examples{
# Process AnnotatedPlainTextDocument (example available in NLP package)
doc <- readRDS(system.file("texts", "stanford.rds", package = "NLP"))
tab <- get_dbpedia_uris(x = doc, language = "en")

tab <- get_dbpedia_uris(
  x = doc,
  language = "en",
  api = "http://api.dbpedia-spotlight.org/en/annotate"
)

tab
library(polmineR)
use("RcppCWB")

uritab <- corpus("REUTERS") \%>\% 
  subset(id == "127") \%>\%
  get_dbpedia_uris(language = "en", p_attribute = "word")

use("GermaParl2")

uritab2 <- corpus("GERMAPARL2MINI") \%>\% 
  subset(speaker_name == "Carlo Schmid") \%>\%
  subset(p_type == "speech") \%>\% 
  get_dbpedia_uris(language = "de", s_attribute = "ne", max_len = 5067)
  
uritab <- corpus("REUTERS") \%>\% 
  split(s_attribute = "id", verbose = FALSE) \%>\% 
  get_dbpedia_uris(language = "en", p_attribute = "word", verbose = TRUE)
}
