% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Cut long string into overlapping segments}
\usage{
segment(x, max_len = 7900L, overlap = 500L)
}
\arguments{
\item{x}{A string (length-one character vector) to process.}

\item{max_len}{The maximum length of (URL-escaped!) string that can be
processed.}

\item{overlap}{Number of overlapping characters.}
}
\value{
A named character vector. The names are integer numbers that indicate
the character offset from the original string.
}
\description{
Strings that are too long to be processed by DBpedia Spotlight are cut into
overlapping segments that can be processed. Overlaps ensure that contextual
information is available for all entities.
}
\examples{
library(polmineR)
use("RcppCWB") # make REUTERS corpus available

article <- corpus("REUTERS") \%>\%
  subset(id == "236") \%>\% # the longest article in the REUTERS corpus
  get_token_stream(p_attribute = "word", collapse = " ")

segs <- segment(x = article, max_len = 500, overlap = 100)
}
