% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{dbpedia_get_wikidata_uris}
\alias{dbpedia_get_wikidata_uris}
\alias{add_wikidata_uris}
\alias{add_wikidata_uris,data.table-method}
\title{Get Wikipedia IDs for DBpedia IDS.}
\usage{
dbpedia_get_wikidata_uris(
  x,
  optional,
  endpoint,
  limit = 100,
  wait = 1,
  verbose = TRUE,
  progress = FALSE
)

add_wikidata_uris(x, ...)

\S4method{add_wikidata_uris}{data.table}(
  x,
  optional,
  endpoint,
  limit = 100,
  wait = 1,
  verbose = TRUE,
  progress = FALSE
)
}
\arguments{
\item{x}{A character vector with DBpedia URIs. NA values are dropped, only
unique values will be processed.}

\item{optional}{Optional information to retrieve (passed as length-one
character vector, e.g. 'municipalityCode').}

\item{endpoint}{Endpoint to query (a \code{character} vector).}

\item{limit}{Single numeric value with maximum size of chunks to process at
a time.}

\item{wait}{A numeric value passed into \code{Sys.sleep()} to slow down sequence
of requests (and avoid denial of service). Defaults to 100.}

\item{verbose}{Whether to show messages (\code{logical} value).}

\item{progress}{Whether to show progress bar (\code{logical} value).}

\item{...}{Further arguments.}
}
\description{
Get Wikipedia IDs for DBpedia IDS.
}
\examples{
dbpedia_uris <- c(
  "http://de.dbpedia.org/resource/Killesberg",
  "http://de.dbpedia.org/resource/Ljubljana",
  "http://de.dbpedia.org/resource/Velbert"
)
dbpedia_get_wikidata_uris(
  dbpedia_uris,
  optional = "municipalityCode",
  endpoint = "http://de.dbpedia.org/sparql",
  wait = 0,
  limit = 2,
  progress = TRUE
)
library(dbpedia)
library(quanteda)
library(dplyr)
library(tibble)

options(dbpedia.lang = "en")
options(dbpedia.endpoint = "http://api.dbpedia-spotlight.org/en/annotate")

httr::set_config(httr::config(ssl_verifypeer = 0L))

uritab <- data_char_ukimmig2010 |>
  corpus() |>
  get_dbpedia_uris() \%>\% 
  add_wikidata_uris(endpoint = "https://dbpedia.org/sparql/", progress = TRUE, limit = 50) \%>\% 
  wikidata_query(id = "P31") \%>\% 
  as_tibble() \%>\% 
  filter(grepl("political party", P31))
  
}
