% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_enrich}
\alias{xml_enrich}
\title{Add entities to original XML}
\usage{
xml_enrich(
  xml,
  annotation_dt,
  entity_tags = c("w", "pc", "name"),
  entity_name = "name",
  token_tags = c("w", "pc"),
  attributes = "dbpedia_uri"
)
}
\arguments{
\item{xml}{...}

\item{annotation_dt}{...}

\item{entity_tags}{...}

\item{entity_name}{...}

\item{token_tags}{...}

\item{attributes}{...}
}
\description{
Add entities to original XML
}
