% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_enrich}
\alias{xml_enrich}
\title{Add entities to original XML}
\usage{
xml_enrich(
  xml,
  annotation_dt,
  token_tags = c("w", "pc"),
  entity_name = "name",
  feature_tag = NULL,
  ref = "dbpedia_uri",
  type = NULL
)
}
\arguments{
\item{xml}{The XML document to be modified in place.}

\item{annotation_dt}{A \code{data.frame} or \code{data.table} of annotations returned
by \code{get_dbpedia_uris()}. Must contain token IDs and values to be added
(e.g. URIs).}

\item{token_tags}{A \verb{character vector}, the names of XML nodes containing
tokens, i.e. which potentially can be (part of) entities.}

\item{entity_name}{A \verb{character vector}, the name of added XML nodes
describing entities.}

\item{feature_tag}{A \verb{character vector}, the name of pre-annotated features
to be used and enriched.}

\item{ref}{A \verb{character vector}, the name of the URI to be added.
Must correspond to a column name in \code{annotation_dt}.}

\item{type}{A \verb{character vector}, the name of the entity type to be added.
Must correspond to a column name in \code{annotation_dt}.}
}
\description{
Enrich original XML with retrieved Uniform Resource Identifiers and
additional information.
}
\details{
If feature_tag is not NULL, then existing entities are enriched when they
entirely overlap with new annotations. In this case, found \code{type}s are added
to the node. If they are not identical, new types are added to previously
annotated types with a pipe, indicating different annotation results.
Regardless of the value in \code{feature_tag}, all annotations in the annotation
data.table are added. To limit the annotation to entities which correspond to
pre-annotated entities, consult the documentation of \code{get_dbpedia_uris()}.
}
